+++ src/const.js
@@ -3,6 +3,9 @@ export const effect_flat_footed_start_turn = "Compendium.pf2e-action-support.act
 export const effect_enfeebled1_start_turn = "Compendium.pf2e-action-support.action-support.Item.9QvJm6xIKWjti3PD";
 export const effect_enfeebled2_start_turn = "Compendium.pf2e-action-support.action-support.Item.66Sg7uVVaIb3tOib";
 export const effect_immobilized1_round = "Compendium.pf2e-action-support.action-support.Item.FCKm1dZKZzF7Jkxl";
+export const effect_blinded1_round = "Compendium.pf2e-action-support.action-support.Item.KmUCaFZSk9efn9AR";
+export const effect_skunk_bomb_fail = "Compendium.pf2e-action-support.action-support.Item.Ed57eZ72j6RGAIp4";
+export const effect_skunk_bomb_cfail = "Compendium.pf2e-action-support.action-support.Item.0v7THqjruKWI0wJ4";
 export const effect_battle_medicine_immunity_hour = "Compendium.pf2e-action-support.action-support.Item.GMb4x4eHVGD9Tfzp";
 export const effect_treat_wounds_immunity_minutes = "Compendium.pf2e-action-support.action-support.Item.b1ILZ8YQvBd9XA2b";
 export const effect_demoralize_immunity_minutes = "Compendium.pf2e-action-support.action-support.Item.DFLW2gzu0PGeX6zu";
@@ -38,6 +41,55 @@ export const spellEffectMap = {
     'armor-of-bones': ['Compendium.pf2e.spell-effects.Item.6BjslHgY01cNbKp5'],
     'athletic-rush': ['Compendium.pf2e.spell-effects.Item.57lnrCzGUcNUBP2O'],
     'true-strike': ['Compendium.pf2e.spell-effects.Item.fpGDAz2v5PG0zUSl'],
+}
+export const featEffectMap = {
+    "smite-evil":"Compendium.pf2e.feat-effects.Item.AlnxieIRjqNqsdVu",
+    "heavens-thunder":"Compendium.pf2e.feat-effects.Item.L9g3EMCT3imX650b",
+    "intimidating-strike":effect_intimidating_strike,
+    "reach-spell":effect_reach_spell,
+
+    "crane-stance":"Compendium.pf2e.feat-effects.Item.nwkYZs6YwXYAJ4ps",
+    "dragon-stance":"Compendium.pf2e.feat-effects.Item.qUowHpn79Dpt1hVn",
+    "gorilla-stance":"Compendium.pf2e.feat-effects.Item.RozqjLocahvQWERr",
+    "monastic-archer-stance":"Compendium.pf2e.feat-effects.Item.1dxD3xsTzak6GNj5",
+    "mountain-stance":"Compendium.pf2e.feat-effects.Item.gYpy9XBPScIlY93p",
+    "rain-of-embers-stance":"Compendium.pf2e.feat-effects.Item.Im5JBInybWFbHEYS",
+    "tiger-stance":"Compendium.pf2e.feat-effects.Item.pf9yvKNg6jZLrE30",
+    "wolf-stance":"Compendium.pf2e.feat-effects.Item.b2kWJuCPj1rDMdwz",
+    "reflective-ripple-stance":"Compendium.pf2e.feat-effects.Item.QDQwHxNowRwzUx9R",
+    "stoked-flame-stance":"Compendium.pf2e.feat-effects.Item.rp1YauUSULuqW8rs",
+    "stumbling-stance":"Compendium.pf2e.feat-effects.Item.BCyGDKcplkJiSAKJ",
+    "shooting-stars-stance":"Compendium.pf2e.feat-effects.Item.RXbfq6oqzVnW6xOV",
+    "cobra-stance":"Compendium.pf2e.feat-effects.Item.CgxYa0lrLUjS2ZhI",
+    "peafowl-stance":"Compendium.pf2e.feat-effects.Item.vjvcccAwdkOLA1Fc",
+    "ironblood":"Compendium.pf2e.feat-effects.Item.tPKXLtDJ3bzJcXlv",
+    "jellyfish-stance":"Compendium.pf2e.feat-effects.Item.pkcr9w5x6bKzl3om",
+    "tangled-forest-stance":"Compendium.pf2e.feat-effects.Item.PMHwCrnh9W4sMu5b",
+    "whirling-blade-stance":"Compendium.pf2e.feat-effects.Item.JefXqvhzUeBArkAP",
+    "vitality-manipulating-stance":"Compendium.pf2e.feat-effects.Item.h45sUZFs5jhuQdCE",
+
+    "everstand-stance": "Compendium.pf2e.feat-effects.Item.GGebXpRPyONZB3eS",
+    "impassable-wall-stance": "Compendium.pf2e.feat-effects.Item.zzC2qZwEKf4Ja3xD",
+
+    "point-blank-shot": "Compendium.pf2e.feat-effects.Item.9HPxAKpP3WJmICBx",
+    "disarming-stance": "Compendium.pf2e.feat-effects.Item.LxSev4GNKv26DbZw",
+    "ricochet-stance-fighter": "Compendium.pf2e.feat-effects.Item.Unfl4QQURWaX2zfd",
+    "mobile-shot-stance": "Compendium.pf2e.feat-effects.Item.NWOmJ6WJFheaGhho",
+    "disruptive-stance": "Compendium.pf2e.feat-effects.Item.qBR3kqGCeKp3T2Be",
+    "dueling-dance-fighter": "Compendium.pf2e.feat-effects.Item.1nCwQErK6hpkNvfw",
+    "lunging-stance": "Compendium.pf2e.feat-effects.Item.W8CKuADdbODpBh6O",
+    "paragons-guard": "Compendium.pf2e.feat-effects.Item.6EDoy3OSFZ4L3Vs7",
+    "graceful-poise": "Compendium.pf2e.feat-effects.Item.mark4VEQoynfYNBF",
+    "multishot-stance": "Compendium.pf2e.feat-effects.Item.l4QUaedYofnfXig0",
+    "twinned-defense-fighter": "Compendium.pf2e.feat-effects.Item.3eHMqVx30JGiJqtM",
+
+    "fanes-fourberie": "Compendium.pf2e.feat-effects.Item.GvqB4M8LrHpzYEvl",
+    "masquerade-of-seasons-stance": "Compendium.pf2e.feat-effects.Item.6IsZQpwRJQWIzdGx",
+    "ricochet-stance-rogue": "Compendium.pf2e.feat-effects.Item.Unfl4QQURWaX2zfd",
+
+    "buckler-dance": "Compendium.pf2e.feat-effects.Item.PS17dsXkTdQmOv7w",
+    "dueling-dance-swashbuckler": "Compendium.pf2e.feat-effects.Item.1nCwQErK6hpkNvfw",
+    "twinned-defense-swashbuckler": "Compendium.pf2e.feat-effects.Item.3eHMqVx30JGiJqtM",
 }
 
 export const actionEffectMap = {